xrRender: Blenders: BmmD separated
